+++ packages/kernel/src/modules/index.ts
@@ -3,6 +3,8 @@ export * from "./manifest/ModuleManifest";
 export * from "./manifest/ModuleDescriptor";
 export * from "./contracts/Module";
 export * from "./contracts/ModuleContext";
+export * from "./contracts/StoppableModule";
+export * from "./contracts/DisposableModule";
 export * from "./default/ModuleRegistry";
 export * from "./default/DefaultModuleLoader";
 export * from "./default/DependencyGraph";